훈련용 목소리 카테고리 CRUD

Closes S03P23C206-94, S03P23C206-95, S03P23C206-96, and S03P23C206-97

See merge request s03-ai-sub3/s03p23c206!49

+++ backend/service_center/views.py
@@ -40,6 +40,9 @@ class QnADetail(APIView):
 
     def delete(self, request, pk):
         qna = QnA.objects.get(pk=pk)
+        if qna.is_answer:
+            return Response("답변이 달린글은 삭제 불가능 합니다.", status=status.HTTP_400_BAD_REQUEST) 
+
         if request.user.is_staff or request.user == qna.user:
             qna.delete()
             return Response("삭제완료", status=status.HTTP_200_OK)
